+++ middleware_check_HMAC_test.go
@@ -90,7 +90,7 @@ func getHMACAuthChain(spec APISpec) http.Handler {
 	return chain
 }
 
-func TestHMACAuthSession(t *testing.T) {
+func TestHMACAuthSessionPass(t *testing.T) {
 	spec := createDefinitionFromString(HMACAuthDef)
 	redisStore := RedisStorageManager{KeyPrefix: "apikey-"}
 	healthStore := &RedisStorageManager{KeyPrefix: "apihealth."}
@@ -115,8 +115,8 @@ func TestHMACAuthSession(t *testing.T) {
 	// Prep the signature string
 	tim := time.Now().Format(refDate)
 	req.Header.Add("Date", tim)
-	signatureString := strings.ToLower("Date") + ":" + url.QueryEscape(tim)
-	log.Debug("Signature string before encoding: ", signatureString)
+	signatureString := strings.ToLower("Date") + ": " + tim
+	log.Info("[TEST] Signature string before encoding: ", signatureString)
 
 	// Encode it
 	key := []byte(thisSession.HmacSecret)
@@ -126,7 +126,7 @@ func TestHMACAuthSession(t *testing.T) {
 	sigString := base64.StdEncoding.EncodeToString(h.Sum(nil))
 	encodedString := url.QueryEscape(sigString)
 	log.Debug("Encoded signature string: ", encodedString)
-	log.Debug("URL Encoded: ", url.QueryEscape(encodedString))
+	log.Info("[TEST] URL Encoded: ", url.QueryEscape(encodedString))
 
 	log.Debug("Signature string: ", fmt.Sprintf("Signature keyId=\"9876\",algorithm=\"hmac-sha1\",signature=\"%s\"", encodedString))
 
@@ -169,7 +169,7 @@ func TestHMACAuthSessionAuxDateHeader(t *testing.T) {
 	// Prep the signature string
 	tim := time.Now().Format(refDate)
 	req.Header.Add("x-aux-date", tim)
-	signatureString := strings.ToLower("x-aux-date") + ":" + url.QueryEscape(tim)
+	signatureString := strings.ToLower("x-aux-date") + ": " + tim
 	log.Debug("Signature string before encoding: ", signatureString)
 
 	// Encode it
@@ -223,7 +223,7 @@ func TestHMACAuthSessionFailureDateExpired(t *testing.T) {
 	// Prep the signature string
 	tim := time.Now().Format(refDate)
 	req.Header.Add("Date", tim)
-	signatureString := strings.ToLower("Date") + ":" + url.QueryEscape(tim)
+	signatureString := strings.ToLower("Date") + ":" + tim
 	log.Debug("Signature string before encoding: ", signatureString)
 
 	// Encode it
@@ -278,7 +278,7 @@ func TestHMACAuthSessionKeyMissing(t *testing.T) {
 	// Prep the signature string
 	tim := time.Now().Format(refDate)
 	req.Header.Add("Date", tim)
-	signatureString := strings.ToLower("Date") + ":" + url.QueryEscape(tim)
+	signatureString := strings.ToLower("Date") + ":" + tim
 	log.Debug("Signature string before encoding: ", signatureString)
 
 	// Encode it
@@ -308,7 +308,7 @@ func TestHMACAuthSessionKeyMissing(t *testing.T) {
 	}
 }
 
-func TestHMACAuthSessionmalformedHeader(t *testing.T) {
+func TestHMACAuthSessionMalformedHeader(t *testing.T) {
 	spec := createDefinitionFromString(HMACAuthDef)
 	redisStore := RedisStorageManager{KeyPrefix: "apikey-"}
 	healthStore := &RedisStorageManager{KeyPrefix: "apihealth."}
@@ -333,7 +333,7 @@ func TestHMACAuthSessionmalformedHeader(t *testing.T) {
 	// Prep the signature string
 	tim := time.Now().Format(refDate)
 	req.Header.Add("Date", tim)
-	signatureString := strings.ToLower("Date") + ":" + url.QueryEscape(tim)
+	signatureString := strings.ToLower("Date") + ":" + tim
 	log.Debug("Signature string before encoding: ", signatureString)
 
 	// Encode it
@@ -360,5 +360,64 @@ func TestHMACAuthSessionmalformedHeader(t *testing.T) {
 
 	if recorder.Code != 400 {
 		t.Error("Request should have failed with key not found error!: \n", recorder.Code)
+	}
+}
+
+func TestHMACAuthSessionPassWithHeaderField(t *testing.T) {
+	spec := createDefinitionFromString(HMACAuthDef)
+	redisStore := RedisStorageManager{KeyPrefix: "apikey-"}
+	healthStore := &RedisStorageManager{KeyPrefix: "apihealth."}
+	orgStore := &RedisStorageManager{KeyPrefix: "orgKey."}
+	spec.Init(&redisStore, &redisStore, healthStore, orgStore)
+	thisSession := createHMACAuthSession()
+
+	// Basic auth sessions are stored as {org-id}{username}, so we need to append it here when we create the session.
+	spec.SessionManager.UpdateSession("9876", thisSession, 60)
+
+	uri := "/"
+	method := "GET"
+
+	recorder := httptest.NewRecorder()
+	param := make(url.Values)
+	req, err := http.NewRequest(method, uri+param.Encode(), nil)
+
+	refDate := "Mon, 02 Jan 2006 15:04:05 MST"
+
+	// Signature needs to be: Authorization: Signature keyId="hmac-key-1",algorithm="hmac-sha1",signature="Base64(HMAC-SHA1(signing string))"
+
+	// Prep the signature string
+	tim := time.Now().Format(refDate)
+	req.Header.Add("Date", tim)
+	req.Header.Add("X-Test-1", "hello")
+	req.Header.Add("X-Test-2", "world")
+	signatureString := strings.ToLower("(request-target): ") + "get /\n"
+	signatureString += strings.ToLower("Date") + ": " + tim + "\n"
+	signatureString += strings.ToLower("X-Test-1") + ": " + "hello" + "\n"
+	signatureString += strings.ToLower("X-Test-2") + ": " + "world"
+	log.Info("[TEST] Signature string before encoding: ", signatureString)
+
+	// Encode it
+	key := []byte(thisSession.HmacSecret)
+	h := hmac.New(sha1.New, key)
+	h.Write([]byte(signatureString))
+
+	sigString := base64.StdEncoding.EncodeToString(h.Sum(nil))
+	encodedString := url.QueryEscape(sigString)
+	log.Debug("Encoded signature string: ", encodedString)
+	log.Info("[TEST] URL Encoded: ", url.QueryEscape(encodedString))
+
+	log.Debug("Signature string: ", fmt.Sprintf("Signature keyId=\"9876\",algorithm=\"hmac-sha1\",signature=\"%s\"", encodedString))
+
+	req.Header.Add("Authorization", fmt.Sprintf("Signature keyId=\"9876\",algorithm=\"hmac-sha1\",headers=\"(request-target) date x-test-1 x-test-2\",signature=\"%s\"", encodedString))
+
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	chain := getHMACAuthChain(spec)
+	chain.ServeHTTP(recorder, req)
+
+	if recorder.Code != 200 {
+		t.Error("Initial request failed with non-200 code, should have gone through!: \n", recorder.Code)
 	}
 }